+++ rest-api/src/main/java/io/nobt/rest/payloads/SimpleViolation.java
@@ -7,15 +7,11 @@ public class SimpleViolation {
     @JsonProperty("property")
     private final String property;
 
-    @JsonProperty("value")
-    private final String value;
+    @JsonProperty("reason")
+    private final String reason;
 
-    @JsonProperty("message")
-    private final String message;
-
-    public SimpleViolation(String property, String value, String message) {
+    public SimpleViolation(String property, String reason) {
         this.property = property;
-        this.value = value;
-        this.message = message;
+        this.reason = reason;
     }
 }